\docType{package}
\name{Robocoap-package}
\alias{Robocoap-package}
\title{Robocoap}
\description{
Generation of co-apperance matrices for a text in txt and pdf format
along with utilities to aid the generation of Gephi diagrams.

Currently only novels format is analyzed, future versions will handle
theather plays and research papers.
}
\details{
  \tabular{ll}{ Package: \tab Robocoap\cr Type: \tab Package\cr
    License:\tab GPL3\cr }
}

\seealso{
  Useful links:
  \itemize{
    \item Main Repository at \url{https://github.com/viraltux/Robocoap}
    \item Robocoap website at \url{https://viraltux.github.io/Robocoap}
    \item Report bugs at
    \url{https://github.com/viraltux/Robocoap/issues}
    \item Gephi website at \url{https://gephi.org}
    }
  }
    
\author{
  \strong{Maintainer}: Fran Urbano \email{viraltux@gmail.com} [copyright holder]
}